pyproject.toml
[project]
name = "thundersorter"
version = "0.1.0"
description = "Local service that classifies emails using Google Gemini for Thunderbird auto-tagging"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "google-genai>=1.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
]

[tool.setuptools.packages.find]
include = ["service*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "unit: unit tests",
    "integration: integration tests",
]
